# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Naiad API
  version: "1.0.0"
  description: Garden irrigation controller — REST API
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LoginRequest:
      type: object
      required: [password]
      properties:
        password:     { type: string }
        device_label: { type: string, description: "Optional label shown in token list" }
    LoginResponse:
      type: object
      required: [token, expires_at]
      properties:
        token:      { type: string }
        expires_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Authenticate with password, receive Bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/auto-login:
    post:
      operationId: autoLogin
      summary: Auto-login for trusted embedding contexts (?embed=1)
      security: []
      description: |
        Issues a token when all configured trigger conditions match:
        Referer in trusted_referers AND (optionally) client IP in trusted_ips.
        The client must assert that ?embed=1 was present in the URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [embed_param_present]
              properties:
                embed_param_present: { type: boolean }
                device_label:        { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '403':
          description: Trigger conditions not met
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/verify:
    get:
      operationId: verifyToken
      summary: Check if current Bearer token is valid
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        '401':
          description: Token invalid or expired
...
  /auth/tokens/{token_prefix}:
    delete:
      operationId: revokeToken
      summary: Revoke a token by its prefix
      parameters:
        - name: token_prefix
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Token revoked }
        '404':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── Status ──────────────────────────────────────────────────────────────────
...